search/serializers.py
from rest_framework import serializers
from search.models import Course

class CourseSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Course
        fields = ('title', 'description')
